refactor update to updateDashboardState and redo tests
