--- packages/dashboard/src/dashboard-actions/update.ts
@@ -1,13 +1,0 @@
-import { DashboardState } from '../types';
-import deepMerge from 'deepmerge';
-
-/**
- * Returns updated dashboard configuration
- */
-export const update = ({
-  state,
-  fieldsToUpdate,
-}: {
-  state: DashboardState;
-  fieldsToUpdate: Partial<DashboardState>;
-}): DashboardState => deepMerge({ ...state }, fieldsToUpdate);
